nkpk: Add support for firmware updates
This patch moves the nk3 update command to the trussed module and
enables it for the nkpk command.
