+++ pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
   "intelhex >=2.3, <3",
   "libusb1 >=3, <4",
   "nethsm >=2.0.1, <3",
-  "nitrokey >=0.4, <0.5",
+  "nitrokey >=0.4.2, <0.5",
   "nkdfu >=0.2, <0.3",
   "pyusb >=1.2, <2",
   "pyserial >=3.5, <4",
